refactor: reduce empty cycling deps (#35570)
* chore: rm in root

* chore: fix ts

* test: Update snapshot

* chore: ignore part
