+++ components/config-provider/context.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
-import type { RenderEmptyHandler } from './renderEmpty';
-import defaultRenderEmpty from './renderEmpty';
+import type { RenderEmptyHandler } from './defaultRenderEmpty';
 import type { Locale } from '../locale-provider';
 import type { SizeType } from './SizeContext';
 import type { RequiredMark } from '../form/Form';
@@ -26,7 +25,7 @@ export interface ConfigConsumerProps {
   rootPrefixCls?: string;
   iconPrefixCls?: string;
   getPrefixCls: (suffixCls?: string, customizePrefixCls?: string) => string;
-  renderEmpty: RenderEmptyHandler;
+  renderEmpty?: RenderEmptyHandler;
   csp?: CSPConfig;
   autoInsertSpaceInButton?: boolean;
   input?: {
@@ -54,11 +53,10 @@ const defaultGetPrefixCls = (suffixCls?: string, customizePrefixCls?: string) =>
   return suffixCls ? `ant-${suffixCls}` : 'ant';
 };
 
+// zombieJ: 🚨 Do not pass `defaultRenderEmpty` here since it will case circular dependency.
 export const ConfigContext = React.createContext<ConfigConsumerProps>({
   // We provide a default function for Context without provider
   getPrefixCls: defaultGetPrefixCls,
-
-  renderEmpty: defaultRenderEmpty,
 });
 
 export const ConfigConsumer = ConfigContext.Consumer;
