+++ components/config-provider/defaultRenderEmpty.tsx
@@ -3,7 +3,7 @@ import Empty from '../empty';
 import type { ConfigConsumerProps } from '.';
 import { ConfigConsumer } from '.';
 
-const renderEmpty = (componentName?: string): React.ReactNode => (
+const defaultRenderEmpty = (componentName?: string): React.ReactNode => (
   <ConfigConsumer>
     {({ getPrefixCls }: ConfigConsumerProps) => {
       const prefix = getPrefixCls('empty');
@@ -19,13 +19,16 @@ const renderEmpty = (componentName?: string): React.ReactNode => (
         case 'Transfer':
         case 'Mentions':
           return <Empty image={Empty.PRESENTED_IMAGE_SIMPLE} className={`${prefix}-small`} />;
+
+        /* istanbul ignore next */
         default:
+          // Should never hit if we take all the component into consider.
           return <Empty />;
       }
     }}
   </ConfigConsumer>
 );
 
-export type RenderEmptyHandler = typeof renderEmpty;
+export type RenderEmptyHandler = typeof defaultRenderEmpty;
 
-export default renderEmpty;
+export default defaultRenderEmpty;
