+++ components/list/index.tsx
@@ -11,6 +11,7 @@ import type { PaginationConfig } from '../pagination';
 import Pagination from '../pagination';
 import { Row } from '../grid';
 import Item from './Item';
+import defaultRenderEmpty from '../config-provider/defaultRenderEmpty';
 
 export { ListItemProps, ListItemMetaProps } from './Item';
 
@@ -259,7 +260,7 @@ function List<T>({
       <ul className={`${prefixCls}-items`}>{items}</ul>
     );
   } else if (!children && !isLoading) {
-    childrenContent = renderEmptyFunc(prefixCls, renderEmpty);
+    childrenContent = renderEmptyFunc(prefixCls, renderEmpty || defaultRenderEmpty);
   }
 
   const paginationPosition = paginationProps.position || 'bottom';
