+++ components/mentions/index.tsx
@@ -8,6 +8,7 @@ import { ConfigContext } from '../config-provider';
 import { FormItemInputContext } from '../form/context';
 import type { InputStatus } from '../_util/statusUtils';
 import { getMergedStatus, getStatusClassNames } from '../_util/statusUtils';
+import defaultRenderEmpty from '../config-provider/defaultRenderEmpty';
 
 export const { Option } = RcMentions;
 
@@ -93,7 +94,7 @@ const InternalMentions: React.ForwardRefRenderFunction<unknown, MentionProps> = 
       return notFoundContent;
     }
 
-    return renderEmpty('Select');
+    return (renderEmpty || defaultRenderEmpty)('Select');
   };
 
   const getOptions = () => {
